we cannot use `npm start` directly because it turns out that will fork a subprocess in which to run the start script.
because Strider needs to know the PID of the server process in order to kill it later, this doesn't work.

instead, we parse the package.json ourselves and read the value of the start script. then we execute
that process directly in order that we know the PID.

this means we can always kill the server process.